requirements.txt
notion-client>=2.0.0
requests>=2.25.1
python-dotenv>=0.19.0
SQLAlchemy>=1.4.0
pytest>=6.2.5
